Replaced moment inversion with adaptive Wheeler algorithm for R+.
Given a vector of 2N + 1 moments, the momentsToZetaWheeler function
calculates the zeta_k quantities to determine the realizability of
the moment vector. It also computes the coefficients of the
recurrence relationship used to define the Jacobi matrix.

The number of base quadrature nodes (QMOM) is adapted based on
the realizability of the moment vector.

The WheelerGQMOMRPlus method receives the output of
momentToZetaWheeler and, if the generalized option is true,
extends the sequence of zeta_k. Two options are available: a
gamma (default) and lognormal extension. The recurrence relation
is also extended to enable the calculation of the GQMOM nodes.

Additional variables were introduced in the GQMOM.py file to
simplify method calls.

TODO: The flux calcuation should be done using quadrature for
      for consistency with the actual reconstruction of the
      distribution provided by GQMOM.